Moved the SdxTestClass; started first test
