+++ integration-test/src/main/resources/testsuites/e2e/aws-sdx-recipe-tests.yaml
@@ -2,6 +2,6 @@ name: "aws-sdx-recipe-tests"
 tests:
   - name: "aws sdx e2e tests"
     classes:
-      - name: com.sequenceiq.it.cloudbreak.testcase.e2e.aws.AwsSdxTests
+      - name: com.sequenceiq.it.cloudbreak.testcase.e2e.sdx.AwsSdxTests
         includedMethods:
           - testSDXMultiRepairIDBRokerAndMasterWithRecipeFile
