+++ integration-test/src/main/resources/testsuites/e2e/aws-sdx-tests.yaml
@@ -2,4 +2,4 @@ name: "aws-sdx-tests"
 tests:
   - name: "aws sdx e2e tests"
     classes:
-      - name: com.sequenceiq.it.cloudbreak.testcase.e2e.aws.AwsSdxTests
+      - name: com.sequenceiq.it.cloudbreak.testcase.e2e.sdx.AwsSdxTests
